Started Transaction class and CSVExtraction logic
