+++ Transactions.cs
@@ -1,0 +1,43 @@
+// using System.Data.Common;
+// using System.Transactions;
+
+// private class Transactions
+// {
+//     public int id { get; }
+//     public string transactionDate { get; }
+//     public string paidFrom { get; }
+//     public string paidTo { get; }
+//     public string narrative { get; }
+//     public int amountInPence { get; }
+
+//     public Transaction(
+//         int id,
+//         string transactionDate,
+//         string paidFrom,
+//         string paidTo,
+//         string narrative,
+//         int amountInPence,
+//     )
+// }
+
+// public class Dog
+// {
+// //properties below
+// public string Name { get; }
+// public int Age { get; }
+// //constructor below
+// public Dog(string name)
+// {
+// Name = name;
+// Age = 0;
+// }
+// //(object) method below
+// public void Bark()
+// {
+// Console.WriteLine("Woof");
+// }
+// }
+
+// Dog osha = new Dog("Osha");
+// osha.Bark();
+// var ageOfOsha = osha.Age;
